+++ src/sisyphus_mirror/checks.py
@@ -1,12 +1,12 @@
-from re import match
+import re
 from typing import Any
 
-from sisyphus_mirror.regexp import RSYNC_RATE_LIMIT_RE
+RSYNC_RATE_LIMIT_RE = re.compile(r"^\d+(?:(?:\.\d+)?m)?$")
 
 
 def is_rsync_rate_limit(value: Any) -> bool:
     return (
         (isinstance(value, int) and value >= 0)
         or
-        (isinstance(value, str) and bool(match(RSYNC_RATE_LIMIT_RE, value)))
+        (isinstance(value, str) and bool(re.match(RSYNC_RATE_LIMIT_RE, value)))
     )