+++ src/sisyphus_mirror/regexp.py
@@ -1,3 +1,0 @@
-import re
-
-RSYNC_RATE_LIMIT_RE = re.compile(r"^\d+(?:(?:\.\d+)?m)?$")