+++ CHANGELOG.rst
@@ -3,7 +3,19 @@ Changelog
 Unreleased
 ==========
 
-[1.0.0] - 2025-11-04
+[1.1.0] - 2025-12-06
+====================
+Added
+-----
+* Automatic replacement configuration section `[sisyphus-mirror]` to `[sisyphus_mirror]`
+  for configuration loading.
+
+Changed
+-------
+* Default configuration file path changed to `/etc/sisyphus-mirror/sisyphus-mirror.toml`
+  instead of `/etc/sisyphus-mirror/default.toml`.
+
+[1.0.0] - 2025-12-04
 ====================
 Added
 -----